Bug fix in exchange operator

+++ Project/Grasp.cpp
@@ -19,8 +19,7 @@ Solution GRASP::execute(){
 
     show_log( "Iterations:\n", 1);
     for (int i = 0; i < this->iterations; i++){
-        show_log( std::to_string( i+1 ) + " of " + std::to_string( this->iterations ) +
-            " " + std::to_string( calcule_percentage( i, this->iterations ) ) + " %\n", 1 );
+        show_log( std::to_string( i+1 ) + " of " + std::to_string( this->iterations ) + " " + std::to_string( (int) calcule_percentage( i, this->iterations ) ) + " %\n", 1 );
         this->unused_vertices.clear();
         actual = this->solution_generation->random_greedy_generation( Instance::instance()->get_path_vertices() );
         actual.lock_checker();
